se ajustan rutas y encarpetados
