+++ package.json
@@ -16,6 +16,7 @@
     "lucide-react": "^0.487.0",
     "next": "15.1.7",
     "react": "^19.0.0",
+    "react-countup": "^6.5.3",
     "react-dom": "^19.0.0",
     "react-icons": "^5.4.0",
     "swiper": "^11.2.6"
