+++ src/app/modules/home/components/carouselCompanies/CarouselWrapper.tsx
