+++ src/app/modules/home/components/carouselPlaces/CarouselPlaces.tsx
@@ -1,9 +1,9 @@
 import React, { useState } from 'react';
-import { places } from '../../data/places.js'; 
+import { places } from '../../../../data/places'; 
 import CarouselWrapper from './CarouselWrapper';
-import CardPlaces from '../cards/CardPlaces';
+import CardPlaces from '@/app/components/cards/CardPlaces';
+import { Categoria } from '@/app/modules/home/utils/enums/categories';
 import NavCategories from '../navs/NavCategories';
-import { Categoria } from '../../enums/categories';
 
 const CarouselPlaces = () => {
   const [selectedCategory, setSelectedCategory] = useState<Categoria>(Categoria.HISTORIA_CULTURA);
@@ -58,25 +58,6 @@ const CarouselPlaces = () => {
           />
         ))}
       </CarouselWrapper>
-      <div className='pt-5 text-center'>
-              {/* <button className="relative inline-flex items-center justify-center mx-auto group p-0.5">
-          <span className="absolute inset-0 bg-gradient-to-r from-lime-400 via-green-400 to-lime-500 rounded-lg p-[4px]" />
-          <span className="relative flex items-center justify-center w-full h-full bg-gray-50 text-lime-500 font-semibold rounded-lg group-hover:bg-transparent group-hover:text-white transition-all duration-300 px-6 py-2">
-            Explorar
-            <ArrowUpRightIcon className="w-5 h-5 ml-3" />
-          </span>
-        </button> */}
-
-        {/* <button className="relative inline-flex items-center justify-center mx-auto group p-0.5">
-          <span className="absolute inset-0 bg-gradient-to-r from-lime-400 via-lime-500 to-lime-600 rounded-lg p-[4px]" />
-          <span className="relative flex items-center justify-center w-full h-full bg-gray-100 text-lime-600 font-semibold rounded-lg group-hover:bg-transparent group-hover:text-white transition-all duration-300 px-6 py-2">
-            Explorar
-            <ArrowUpRightIcon className="w-5 h-5 ml-3" />
-          </span>
-        </button> */}
-
-
-      </div>
     </div>
   );
 };
