+++ src/app/modules/home/components/carouselPlacesRoutes/CarouselWrapper.tsx
@@ -4,7 +4,7 @@ import { Navigation, Pagination, Autoplay } from "swiper/modules";
 import "swiper/css";
 import "swiper/css/navigation";
 import "swiper/css/pagination";
-import "../../../styles/CarouselWrapper.css";
+import "../../../../../styles/CarouselWrapper.css";
 
 interface CarouselWrapperProps {
   children: ReactNode;
@@ -64,19 +64,19 @@ const CarouselWrapper = ({
             spaceBetween: 20
           },
           900: {
-            slidesPerView: 4,
+            slidesPerView: 3,
             spaceBetween: 22
           },
           1024: {
-            slidesPerView: 4,
+            slidesPerView: 3,
             spaceBetween: 24
           },
           1200: {
-            slidesPerView: 5,
+            slidesPerView: 3,
             spaceBetween: 26
           },
           1400: {
-            slidesPerView: 5,
+            slidesPerView: 4,
             spaceBetween: 28
           }
         }}
