+++ src/app/modules/home/utils/enums/categories.ts
