+++ src/app/page.tsx
@@ -1,12 +1,12 @@
 'use client'
 import Hero from "./components/pages/inicio/hero";
-import CarouselCompanies from "./components/carouselCompanies/CarouselCompanies";
-import GalleryVillages from "./components/gallery/GalleryVillage";
-import CarouselFoods from './components/carouselFood/CarouselFoods';
-import GalleryCategories from "./components/gallery/GalleryCategories";
-import CarouselPlaces from "./components/carouselPlaces/CarouselPlaces";
-import CarouselPlacesRoutes from "./components/carouselPlacesRoutes/CarouselPlacesRoutes";
-import HistorySection from "./components/history/HistorySection";
+import CarouselCompanies from "./modules/home/components/carouselCompanies/CarouselCompanies";
+import GalleryVillages from "./modules/home/components/galleries/GalleryVillage";
+import CarouselFoods from './modules/home/components/carouselFood/CarouselFoods';
+import GalleryCategories from "./modules/home/components/galleries/GalleryCategories";
+import CarouselPlaces from "./modules/home/components/carouselPlaces/CarouselPlaces";
+import CarouselPlacesRoutes from "./modules/home/components/carouselPlacesRoutes/CarouselPlacesRoutes";
+import HistorySection from "./modules/home/components/histories/HistorySection";
 
 export default function Home() {
   return (
@@ -21,35 +21,35 @@ export default function Home() {
 
       {/* Carousel PlacesRoutes */}
       <section>
-        <div className="p-5 md:pl-20 md:pr-20 md:pb-10 md:pt-10 bg-gray-100">
+        <div className="p-5 md:pl-20 md:pr-20 md:pb-10 md:pt-10 bg-white">
           <CarouselPlacesRoutes/> 
         </div>
       </section>
 
       {/* Carousel Places */}
       <section>
-        <div className="p-5 md:pl-20 md:pr-20 md:pb-10 md:pt-10 bg-gray-100">
+        <div className="p-5 md:pl-20 md:pr-20 md:pb-10 md:pt-10 bg-white">
           <CarouselPlaces />
         </div>
       </section>
 
       {/* History Section */}
        <section>
-        <div className="p-5 md:pl-20 md:pr-20 md:pb-10 md:pt-10 bg-gray-100">
+        <div className="p-5 md:pl-20 md:pr-20 md:pb-10 md:pt-10 bg-white">
           <HistorySection />
         </div>
       </section>
 
        {/* Gallery Categories */}
       <section>
-        <div className="p-5 md:pl-20 md:pr-20 md:pb-10 md:pt-10 bg-gray-100">
+        <div className="p-5 md:pl-20 md:pr-20 md:pb-10 md:pt-10 bg-white">
           <GalleryCategories />
         </div>
       </section>
 
       {/* Carousel Foods Definidos en el componente pare tener el efecto */}
       <section>
-      <div className="p-5 md:pl-20 md:pr-20 md:pb-10 md:pt-10 bg-gray-100">
+      <div className="p-5 md:pl-20 md:pr-20 md:pb-10 md:pt-10 bg-white">
         <CarouselFoods />
         </div>
       </section>
@@ -57,7 +57,7 @@ export default function Home() {
 
        {/* Gallery Villages */}
       <section>
-        <div className="p-5 md:pl-20 md:pr-20 md:pb-10 md:pt-10 bg-gray-100">
+        <div className="p-5 md:pl-20 md:pr-20 md:pb-10 md:pt-10 bg-white">
           <GalleryVillages />
         </div>
       </section>
@@ -65,7 +65,7 @@ export default function Home() {
       
       {/* Carousel Companies Definidos en el componente pare tener el efecto */}
       <section>
-        <div className="md:pb-10 md:pt-10 bg-gray-100">
+        <div className="md:pb-10 md:pt-10 bg-white">
           <CarouselCompanies />
         </div>
       </section>
